+++ examples/echo_client.py
@@ -5,7 +5,7 @@ import json
 import logging
 from pathlib import Path
 import sys
-from typing import Any, Dict, FrozenSet, Literal, Optional, Union
+from typing import Any, Dict, FrozenSet, Literal, NamedTuple, Optional, Set, Union
 
 from omemo.storage import Just, Maybe, Nothing, Storage
 from omemo.types import DeviceInformation, JSONType
@@ -13,6 +13,7 @@ from omemo.types import DeviceInformation, JSONType
 from slixmpp.clientxmpp import ClientXMPP
 from slixmpp.jid import JID  # pylint: disable=no-name-in-module
 from slixmpp.plugins import register_plugin  # type: ignore[attr-defined]
+from slixmpp.plugins.xep_0045 import XEP_0045  # type: ignore[attr-defined]
 from slixmpp.stanza import Message
 from slixmpp.xmlstream.handler import CoroutineCallback
 from slixmpp.xmlstream.matcher import MatchXPath
@@ -70,7 +71,7 @@ class XEP_0384Impl(XEP_0384):  # pylint: disable=invalid-name
         "json_file_path": None
     }
 
-    def __init__(self, *args: Any, **kwargs: Any) -> None:  # pylint: disable=redefined-outer-name
+    def __init__(self, *args: Any, **kwargs: Any) -> None:
         super().__init__(*args, **kwargs)
 
         # Just the type definition here
@@ -127,15 +128,23 @@ class XEP_0384Impl(XEP_0384):  # pylint: disable=invalid-name
 register_plugin(XEP_0384Impl)
 
 
+class MUCJoinInfo(NamedTuple):
+    room_jid: JID
+    nick: str
+
+
 class OmemoEchoClient(ClientXMPP):
     """
     A simple Slixmpp bot that will echo encrypted messages it receives.
+    Works in one-to-one chats as well as in MUCs.
 
     For details on how to build a client with Slixmpp, look at examples in the Slixmpp repository.
     """
 
-    def __init__(self, jid: str, password: str) -> None:
+    def __init__(self, jid: str, password: str, muc_join_info: Optional[MUCJoinInfo]) -> None:
         super().__init__(jid, password)
+
+        self.muc_join_info = muc_join_info
 
         self.add_event_handler("session_start", self.start)
         self.register_handler(CoroutineCallback(
@@ -144,7 +153,7 @@ class OmemoEchoClient(ClientXMPP):
             self.message_handler  # type: ignore[arg-type]
         ))
 
-    def start(self, _event: Any) -> None:
+    async def start(self, _event: Any) -> None:
         """
         Process the session_start event.
 
@@ -156,7 +165,12 @@ class OmemoEchoClient(ClientXMPP):
         """
 
         self.send_presence()
-        self.get_roster()  # type: ignore[no-untyped-call]
+        await self.get_roster()  # type: ignore[no-untyped-call]
+
+        xep_0045: Optional[XEP_0045] = self["xep_0045"]
+        if xep_0045 is not None and self.muc_join_info is not None:
+            # Started as a task as a workaround for https://codeberg.org/poezio/slixmpp/issues/3660
+            asyncio.create_task(xep_0045.join_muc_wait(self.muc_join_info.room_jid, self.muc_join_info.nick))
 
     async def message_handler(self, stanza: Message) -> None:
         """
@@ -168,13 +182,38 @@ class OmemoEchoClient(ClientXMPP):
                 to see how it may be used.
         """
 
+        xep_0045: Optional[XEP_0045] = self["xep_0045"]
         xep_0384: XEP_0384 = self["xep_0384"]
 
-        mto = stanza["from"]
+        mfrom: JID = stanza["from"]
 
         mtype = stanza["type"]
-        if mtype not in { "chat", "normal" }:
+        if mtype not in { "chat", "normal", "groupchat" }:
             return
+
+        is_muc_reflection = False
+        if mtype == "groupchat":
+            if xep_0045 is None:
+                log.warning("Ignoring MUC message while MUC plugin is not loaded.")
+                return
+
+            # Save the sender's nick and modify mfrom to just be the MUC JID
+            mfrom_nick = mfrom.resource
+            mfrom = JID(mfrom.bare)
+
+            # Find the real JID of the sender
+            real_mfrom_str: Optional[str] = xep_0045.get_jid_property(JID(mfrom.bare), mfrom_nick, "jid")
+            if real_mfrom_str is None:
+                self.plain_reply(
+                    mfrom,
+                    mtype,
+                    "This bot cannot be used in semi-anonymous MUCs."
+                )
+                return
+
+            # Check if this message is a MUC reflection
+            if JID(real_mfrom_str) == self.boundjid:
+                is_muc_reflection = True
 
         namespace = xep_0384.is_encrypted(stanza)
         if namespace is None:
@@ -182,8 +221,13 @@ class OmemoEchoClient(ClientXMPP):
                 # This is the case for things like read markers, ignore those.
                 return
 
+            # Don't respond to reflected MUC messages
+            if is_muc_reflection:
+                log.debug(f"Ignoring reflected unencrypted MUC message: {stanza['body']}")
+                return
+
             self.plain_reply(
-                mto,
+                mfrom,
                 mtype,
                 f"Unencrypted message or unsupported message encryption: {stanza['body']}"
             )
@@ -200,11 +244,18 @@ class OmemoEchoClient(ClientXMPP):
                 # This is the case for things like read markers, ignore those.
                 return
 
-            await self.encrypted_reply(mto, mtype, message)
-        except Exception as e:  # pylint: disable=broad-exception-caught
-            self.plain_reply(mto, mtype, f"Error {type(e).__name__}: {e}")
+            # Don't respond to reflected MUC messages. Cool to see though if message reflection in encrypted
+            # MUCs works correctly
+            if is_muc_reflection:
+                log.info(f"Ignoring reflected encrypted MUC message: {message['body']}")
+                return
 
-    def plain_reply(self, mto: JID, mtype: Literal["chat", "normal"], reply: str) -> None:
+            await self.encrypted_reply(mfrom, mtype, message)
+        except Exception as e:  # pylint: disable=broad-exception-caught
+            log.warning("Exception while handling encrypted message", exc_info=True)
+            self.plain_reply(mfrom, mtype, f"Error {type(e).__name__}: {str(e)}")
+
+    def plain_reply(self, mto: JID, mtype: Literal["chat", "normal", "groupchat"], reply: str) -> None:
         """
         Helper to reply with plain messages.
 
@@ -221,7 +272,7 @@ class OmemoEchoClient(ClientXMPP):
     async def encrypted_reply(
         self,
         mto: JID,
-        mtype: Literal["chat", "normal"],
+        mtype: Literal["chat", "normal", "groupchat"],
         reply: Union[Message, str]
     ) -> None:
         """
@@ -243,9 +294,17 @@ class OmemoEchoClient(ClientXMPP):
         reply.set_to(mto)
         reply.set_from(self.boundjid)
 
-        # It might be a good idea to strip everything bot the body from the stanza, since some things might
-        # break when echoed.
-        messages, encryption_errors = await xep_0384.encrypt_message(reply, mto)
+        encrypt_for: Set[JID] = { mto }
+        if mtype == "groupchat":
+            # In a MUC, encrypt for all participants
+            xep_0045: XEP_0045 = self["xep_0045"]
+            encrypt_for = {
+                JID(xep_0045.get_jid_property(mto, nick, "jid"))
+                for nick
+                in xep_0045.get_roster(mto)
+            }
+
+        messages, encryption_errors = await xep_0384.encrypt_message(reply, encrypt_for)
 
         if len(encryption_errors) > 0:
             log.info(f"There were non-critical errors during encryption: {encryption_errors}")
@@ -256,7 +315,7 @@ class OmemoEchoClient(ClientXMPP):
             message.send()
 
 
-if __name__ == "__main__":
+def main() -> None:
     # Set up the command line argument parser
     parser = ArgumentParser(description=OmemoEchoClient.__doc__)
 
@@ -271,7 +330,44 @@ if __name__ == "__main__":
              " shared between multiple accounts."
     )
 
+    parser.add_argument(
+        "-r", "--room-jid",
+        dest="room_jid",
+        help="JID of a MUC to join. --nick must also be set if this is used."
+    )
+
+    parser.add_argument(
+        "-n", "--nick",
+        dest="nick",
+        help="Nickname to use in the MUC specified by --room-jid."
+    )
+
+    parser.add_argument(
+        "-v", "--verbose",
+        dest="log_level",
+        action="count",
+        default=0,
+        help="logging verbosity (WARNING (default), INFO (-v), DEBUG (-vv))"
+    )
+
     args = parser.parse_args()
+
+    logging.basicConfig(level=[
+        logging.WARNING,
+        logging.INFO
+    ][min(args.log_level, 1)])
+
+    log.setLevel([
+        logging.WARNING,
+        logging.INFO,
+        logging.DEBUG
+    ][min(args.log_level, 2)])
+
+    muc_join_info: Optional[MUCJoinInfo] = None
+    if args.room_jid is not None:
+        if args.nick is None:
+            raise ValueError("Please set both the room JID of a MUC to join and the nickname to use.")
+        muc_join_info = MUCJoinInfo(room_jid=JID(args.room_jid), nick=args.nick)
 
     # Ask for username and password if not supplied via cli args
     if args.username is None:
@@ -282,7 +378,8 @@ if __name__ == "__main__":
     # Setup the OmemoEchoClient and register plugins. Note that while plugins may have interdependencies, the
     # order in which you register them does not matter.
 
-    xmpp = OmemoEchoClient(args.username, args.password)
+    xmpp = OmemoEchoClient(args.username, args.password, muc_join_info)
+    xmpp.register_plugin("xep_0045")  # Multi-User Chat
     xmpp.register_plugin("xep_0199")  # XMPP Ping
     xmpp.register_plugin("xep_0380")  # Explicit Message Encryption
     xmpp.register_plugin(
@@ -294,3 +391,7 @@ if __name__ == "__main__":
     # Connect to the XMPP server and start processing XMPP stanzas.
     xmpp.connect()
     asyncio.get_event_loop().run_forever()
+
+
+if __name__ == "__main__":
+    main()